requirements.txt
twilio==8.10.0
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
gunicorn==21.2.0
python-dateutil==2.8.2 